Allow for optional list of bufnames to not resize
usage:

let g:eighties_bufname_additional_patterns = ['fugitiveblame']

This way if you do a :Gblame on a file, then navigate around with the
blame open, it won't get re-sized. It's the same behavior we want from
NERDTree too, but instead of just hard-coding it in that array I made it
configurable.